coverity: fix the Overflowed return value in msr.c
coverity report the error:
Event return_overflow: "status", which might have overflowed, is
returned from the function.
It's because msr_reg_read return u64 while the status return in
process_args is int tpe.

Tests Done:
1. Build and boot
2. peeknpoke function works

Tracked-On: OAM-129214

+++ msr.c
@@ -28,7 +28,7 @@
 /*
  *  This function reads from requested MSR address
  */
-uint64_t msr_reg_read(int cpu, unsigned int reg_offset, int print_enabled)
+int msr_reg_read(int cpu, unsigned int reg_offset, int print_enabled)
 {
 	char buf[128];
 	uint64_t temp = 0;
